% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procdata.R
\docType{data}
\name{procdata}
\alias{procdata}
\title{Sinlge-cell RNA-seq data in TSCAN package after preprocessing}
\format{
A matrix with 534 rows and 131 variables
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE48968}
}
\usage{
procdata
}
\description{
The dataset contains 534 rows and 131 columns. Each row represent a gene and each column represent a single cell.
}
\references{
Shalek, A. K., Satija, R., Shuga, J., Trombetta, J. J., Gennert, D., Lu, D., ... & Regev, A. (2014). Single-cell RNA-seq reveals dynamic paracrine control of cellular variation. Nature.
Ji, Z., & Ji, H. (2016). TSCAN: Pseudo-time reconstruction and evaluation in single-cell RNA-seq analysis. Nucleic acids research, 44(13), e117-e117.
}
\keyword{datasets}
