% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPC.R
\name{findPC}
\alias{findPC}
\title{findPC}
\usage{
findPC(
  sdev,
  number = 30,
  method = "perpendicular line",
  aggregate = NULL,
  figure = FALSE
)
}
\arguments{
\item{sdev}{A vector of the standard deviations explained by each principal component}

\item{number}{A vector including number of PCs used in the function}

\item{method}{Methods include 'all', 'piecewise linear model', 'first derivative', 'second derivative', 'preceding residual', 'perpendicular line (default)', 'k-means clustering'}

\item{aggregate}{An option to report the mean, median, or voting (median if all are different, otherwise mode) of the results}

\item{figure}{Whether to draw a heatmap showing the results}
}
\value{
A numeric value (matrix) including the number of PCs.
}
\description{
Automatically select the number of principal components in single-cell analysis
}
\details{
findPC returns a list of results for one or multiple methods with different total numbers of PCs.
It also supports the aggregation of results from multiple methods and total numbers of PCs by taking the average, median, or mode.
For users who want to visually inspect the results, findPC can also display the elbow plot with the automatically identified elbow points.
}
\examples{
# Return the default result (Perpendicular line with 30 PCs)
findPC(sdev = sdev)

Return the results based on Piecewise linear model with 16,20 and 28 PCs
findPC(sdev = sdev,number = c(16,20,28),method = 'piecewise linear model')

Return the results based on six methods with 16,20 and 28 PCs
findPC(sdev = sdev,number = c(16,20,28),method = 'all')

Return voting (median if all are different, otherwise mode) of the results from six methods with 16,20 and 28 PCs
findPC(sdev = sdev,number = c(16,20,28),method = 'all',aggregate = 'voting')

Return a heatmap based on six methods with 16,20 and 28 PCs
findPC(sdev = sdev,number = c(16,20,28),method = 'all',figure = T)
}
\author{
Haotian Zhuang, Zhicheng Ji \href{mailto:haotian.zhuang@duke.edu}{haotian.zhuang@duke.edu}
}
