% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procdata.R
\docType{data}
\name{procdata}
\alias{procdata}
\title{Single-cell RNA-seq data}
\format{
A matrix with 534 rows and 131 columns
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE48968}
}
\usage{
procdata
}
\description{
The dataset contains 534 rows and 131 columns. Each row represents a gene and each column represents a single cell.
}
\references{
Shalek, A. K., Satija, R., Shuga, J., Trombetta, J. J., Gennert, D., Lu, D., ... & Regev, A. (2014). Single-cell RNA-seq reveals dynamic paracrine control of cellular variation. Nature.
}
\keyword{datasets}
